added extra methods and changed conditional checking for selling games
Implemented more of the uml desing methods into the class and changed conditional checking when selling a game, now the seller is also checked to see if he owns the game in order to sell it and the game also then also removed when selling it to another person. oh and corrected a spelling mistake.